==> Proyecto 2/notebook/Draft/02. CSV_inicial_-_Enriquecido_con_CSV_Selenium_.ipynb <==
# Mergeo de CSV Inicial y CSV creado a través de Selenium

import pandas as pd
import numpy as np
import sidetable

import seaborn as sns
import matplotlib.pyplot as plt

Nos descargamos el CSV que vamos a utilizar de Kaggle.

Los datos son sobre casas en los Estados de Washington, California y Oregón, y divididos por ciudades. Incluye solo algunas de las ciudades más importantes de cada estado.
Los datos incluyen el precio, dirección, número de cuartos, baños, el área de la casa, etc.

Link al kaggle: https://www.kaggle.com/ialjadani/houses-prices-in-selected-area-in-usa

df = pd.read_csv ("../Data/02. CSV_inicial_houses_USA.csv")
df.sort_values(by = ["City"]).head(3)

df.shape

set(df["City"])

A continuación nos descargamos un CSV que hemos creado a partir de Selenium "SeleniumProMo"

df_calidad_de_vida_eeuu_ciudades = pd.read_csv ("Data/01. Resultado_Selenium_datos_ciudades_EEUU.csv")
df_calidad_de_vida_eeuu_ciudades.head()

A continuaición vamos a mergear ambos CSV:

mergeado_inner = df.merge(df_calidad_de_vida_eeuu_ciudades, on= "City")

Exportamos el DataFrame mergeado a un CSV

mergeado_inner.to_csv("03. Datos_mergeados.csv")

# Análisis Exploratorio

import pandas as pd
import numpy as np
import sidetable

import seaborn as sns
import matplotlib.pyplot as plt

mergeado_inner.head(2)

# Exploración inicial de nuestros datos

### Tipado de Datos

mergeado_inner.columns

!!!! OJO !!!! Vemos como los nombres no son homogéneos y eso nos puede dificultar nuestro trabajo. Habrá que ponerkos en minúscula, cambiar espacios por "_", y eliminar las columnas que no nos aporten valor

### Tamaño

mergeado_inner.shape

### Tipos de datos

¿Y el tipo de datos de cada una de las columnas?

mergeado_inner.dtypes

!!!! OJO !!!! No tiene mucho sentido que el número de 'baños' sea un "float", ni que los números de "Índice_poder_adqui', 'Índice_seguridad', 'Índice_sanidad', 'Índice_clima', 'Índice_costo_vida', 'Relación_precio_ingresos_propiedades', 'Índice_costo_desplazamiento', 'Índice_contaminación', 'Índice_calidad_vida' sean un "string"... Tendremos que hacer algo con ellas más adelante.



### Info

mergeado_inner.info()

No tenemos valores nulos en nuestrio DataFrame

### Principales estadísticos

mergeado_inner.describe().T

mergeado_inner.describe(include = "object").T

### Duplicados

¿Tenemos duplicados en nuestro dataframe? -> Parece que no tenemos....

mergeado_inner.duplicated().sum()

Pero, qué está pasando? Vamos a eliminar dos columnas con valores únicos "Unnamed: 0_x", "Unnamed: 0_y", y vamos a comprobar si realmente seguimos sin tener duplicados

eliminar_columns = ['Unnamed: 0_x', 'Unnamed: 0_y']

mergeado_inner_1 = mergeado_inner.drop(eliminar_columns, axis = 1)
mergeado_inner_1.head(2)

mergeado_inner_1.duplicated().sum()

!!!! OJO!!!!  vemos como tenemos dos valores que están duplicados en nuestro dataframe

### Sidetables

mergeado_inner_1.stb.freq(["AD_Situation"])

mergeado_inner_1.stb.freq(["Brokered_by"])

mergeado_inner_1.stb.freq(["City"])

mergeado_inner_1.stb.freq(["State"])

mergeado_inner_1[mergeado_inner_1.duplicated(keep=False)]

### Exportación de CSV mergeado

Exportamos el DataFrame mergeado, y con una limpieza minimísima, a un CSV:

mergeado_inner_1.to_csv("03. Datos_mergeados_.csv")
