==> Proyecto 2/notebook/Draft/03. Método_de_Transformación_-Copy1.ipynb <==
import pandas as pd
import numpy as np
import sidetable
import pymongo

import folium
import seaborn as sns
import matplotlib.pyplot as plt

import sys
sys.path.append("../")
import src.creacion_columnas as cc

### Importamos CSV

Nos descarmos el CSV que hemos creado, enriqueciendo el CSV inicial con Selenium:

df3 = pd.read_csv ("../Data/03. Datos_mergeados_.csv")
df3.head(2)

### Manipulación de columnas

Vamos a reescribir la columna "AD_Situation" para que únicamente muestre dos opciones: New, old. 

Vamos a reescribir la columna "State" para que únicamente muestra la primera parte de los string, es decir: WA, OR o CA.

Para ello, hemos creado una función "quitar_horas" donde haremos un split por espacio y nos quedaremos con la primera parte del string!

"""def quitar_horas(x): 
    try:
        return x.split()[0]
    except: 
        return x"""

df3["AD_Situation"] = df3["AD_Situation"].apply(cc.quitar_horas)
df3["State"] = df3["State"].apply(cc.quitar_horas)
df3.head(1)

De nuevo vamos a reescriir la columna "State", pero para que este vez muestre los 3 Estados en su nomnbre completo: California, Oregón, Washington: 

Para ello hemos creado la función "cambiar"!

""""def cambiar(x):
    if x == "WA":
        return x.replace("WA", "Washington")

    elif  x == "OR":
        return x.replace("OR", "Oregón")

    else:
        return x.replace("CA", "California")"""

df3["State"] = df3.apply(lambda datos: cc.cambiar(datos["State"]),axis = 1)

Vamos a crear una nueva columna "Measuring unit" y vamos a indicar como unidad de medida "m2".

Vamos a reescribir la columna "Bulding space" para que la información se muestra en m2 en vez de sqft. Para ello lo vamos a dividir por 10.764.

Vamos a crear una columna nueva "Price/m2", donde vamos dividir la columna "Price" por la columna "Building space". 

df3['Measuring unit']= df3['Measruing unit'].replace("sqft","m2")
df3['Building space']= round(df3['Building space']/10.764, 0)
df3['Price/m2']= round((df3['Price'])/(df3['Building space']), 0)
df3.head(1)

Vamos a crear tres columnas nuevas, "Direction", "Ciudad1", "Postal_code" a partir de la columna "Address", donde vamos a hacer un split por ",", y nos vamos a quedar con las 3 primeras palabras del string:

df3[["Direction", "Ciudad1", "Postal_code"]]  = df3["Address"].str.split(",",  expand=True).iloc[:, :3]

Vamos a reescribir la columna "Direction" para que nos de la dirección y la ciudad en la misma columna. Para ello, vamos a sumar los strings de las columnas "Direction" y "Ciudad1":

df3["Direction"] = df3["Direction"] + "," + df3["Ciudad1"]
df3.head(1)

Vamos a reescribir la columna "Postal_code" para que nos de únicamente el código postal: 

Para ello, utilizando regex, vamos a quedarnos únicamente con los valores numéricos del string

df3["Postal_code"] = df3.Postal_code.str.extract("(\d+)", expand=True)
df3.head(1)

Vamos a eliminar las columna que no nos interesan por estar la información "duplicada"

eliminar_columns = ['Address', 'Unnamed: 0', 'Ciudad1', 'Measruing unit']


df3_limpio = df3.drop(eliminar_columns, axis = 1)
df3_limpio.head(1)

### Estandarización del tipado de las columnas

Vamos a ordenar las columnas de nuestro DataFrame para que tenga algo más de sentido:

df3_limpio = df3_limpio.reindex(columns=['Price', 'Currency', 'AD_Situation', 'Rooms', 'Bath', 'Building space', 'Measuring unit', 'Price/m2', 'Brokered_by', 'Direction', 'City', 'State', 'Postal_code',
       'Índice_poder_adqui', 'Índice_seguridad', 'Índice_sanidad',
       'Índice_clima', 'Índice_costo_vida',
       'Relación_precio_ingresos_propiedades', 'Índice_costo_desplazamiento',
       'Índice_contaminación', 'Índice_calidad_vida'])
df3_limpio.head(1)

A continuación, vamos a homogeneizar las columnas: pasar los nombres de las columnas a minusculas y reemplazar los espacios por "_":

diccionario = {}
for col in df3_limpio.columns:
    diccionario[col] = col.lower().strip().replace(" ", "_")

nuevas_columnas = {col : col.lower().strip().replace(" ", "_") for col in df3_limpio.columns}
nuevas_columnas

df3_limpio.rename(columns = nuevas_columnas, inplace =True)
df3_limpio.head(1)

### Tipo de Datos

Problema: La columna "baños" está en formato "float". Debemos convertirlas a integer. Las columans "índice_poder_adqui', 'índice_seguridad', 'índice_sanidad', 'índice_clima', 'índice_costo_vida', 'relación_precio_ingresos_propiedades', 'índice_costo_desplazamiento', 'índice_contaminación', 'índice_calidad_vida' están en formato "string"... Tendremos que hacer algo con ellas más adelante. Debemos convertirlas a integer:

df3_limpio.dtypes

df3_limpio["bath"] = df3_limpio["bath"].round(1).astype(int)
df3_limpio.head(2)

x = ['índice_poder_adqui', 'índice_seguridad',
       'índice_sanidad', 'índice_clima', 'índice_costo_vida',
       'relación_precio_ingresos_propiedades', 'índice_costo_desplazamiento',
       'índice_contaminación', 'índice_calidad_vida']

def puntos(df,x):
    for i in x:
        
        df3_limpio[i] = df3_limpio [i].str.replace(",","")
        df3_limpio[i] = pd.to_numeric(df[i], errors = "coerce").astype("Int64")


puntos(df3_limpio, x)

df3_limpio.dtypes

### Duplicados

Problema: Duplicado. En el análisis exploratorio identificamos 2 duplicados:

df3_limpio.duplicated().sum()

df3_limpio[df3_limpio.duplicated() == True]

Una vez identificamos las filas duplicadas, procedemos a su eliminación:

df3_limpio.drop_duplicates(inplace = True)

df3_limpio.duplicated().sum()

### Valores Nulos

df3_limpio.isnull().sum()

df3_limpio["postal_code"] = df3_limpio["postal_code"].fillna(0)

df3_limpio.isnull().sum()

df3_limpio["postal_code"].dtypes

df3_limpio["postal_code"] = df3_limpio["postal_code"].astype(int)

df3_limpio["postal_code"].dtypes

### Principales estadísticos

df3_limpio.describe().T

df3_limpio.describe(include = "object").T

### Geopy

from geopy.geocoders import Nominatim

locator = Nominatim(user_agent= "myGeocoder")

"""location = locator.geocode ("98516")
print (location)"""

"""print("Latitude = {}, Longitude = {}".format(location.latitude, location.longitude))"""

"""location.latitude"""

def latitud_longitud (x):
    locator = Nominatim(user_agent= "myGeocoder")
    try:
        
        location = locator.geocode (x)
        return location.latitude, location.longitude
        
    except:
        return None

latitud_longitud (98516)

"""df3_limpio[["latitud", "longitud"]] = df3_limpio.apply(lambda datos: latitud_longitud(
                                    datos["postal_code"]), axis = 1)"""

"""df3_limpio[["latitud", "longitud"]] = df3_limpio["postal_code"].apply(latitud_longitud).split(",")

def latitud_v2 (x):
    locator = Nominatim(user_agent= "myGeocoder")
    try:
        
        location = locator.geocode (x)
        return location.latitude, location.longitud
        
    except:
        return None

def longitud_v2 (x):
    locator = Nominatim(user_agent= "myGeocoder")
    try:
        
        location = locator.geocode (x)
        return location.longitude
        
    except:
        return None

def hola(x):
    return "Hola", "adios"

x, y = hola()

xñkhj

y

df3_limpio[["laitud", "longtud"]] = df3_limpio["postal_code"].apply(hola)

df3_limpio["latidud_"] = df3_limpio.apply(lambda datos: latitud_v2(
                                    datos["postal_code"]), axis = 1)

df3_limpio["longitud_"] = df3_limpio.apply(lambda datos: longitud_v2(
                                    datos["postal_code"]), axis = 1)

### Exportación de CSV limpio

Exportamos el DataFrame limpio a un CSV:

df3_limpio.to_csv("04. Datos_limpitos_.csv")

"""dd = cc.latitud_v2("98516")"""

dd

mapa=folium.Map(location=[47.134266, -122.734934],  # la posicion
                tiles='openstreetmap',            # tipo de mapa, theme
                zoom_start=10)                    # zoom inicial

mapa

"""for index, row in df.iterrows(): # método para iterar por los dataframes

cities = ['Olympia', 'Seattle', 'Des Moines', 'Bellevue', 'Redmond', 'Tacoma', 'Vancouver', 'Long Beach', 'Pasadena', 'Anaheim', 'Irvine', 'Portland']
    for i in cities:
        try:
                row['city'] == (i) :

        return folium.Marker([row["latitud_"], row['longitud_']],
                              radius=10, 
                              tooltip = row[['is_happy_customer', 'city']],
                              icon=folium.Icon(icon = "cloud", color = "red")).add_to(mapa)
        except:
            None"""

df3_limpio

for index, row in df3_limpio.iterrows():
    try:
        if row['city'] == "Olympia":

            folium.Marker([row["latitud_"], row['longitud_']],
                        radius=10,
                        icon=folium.Icon(icon = "cloud", color = "red")).add_to(mapa)
        elif row['city']  == "Seattle":
            folium.Marker([row["latitud_"], row['longitud_']],
                        radius=10, 
                        icon=folium.Icon(icon = "cloud", color = "orange")).add_to(mapa)

        elif row['city']  == "Des Moines":
            folium.Marker([row["latitud_"], row['longitud_']],
                        radius=10,
                        icon=folium.Icon(icon = "cloud", color = "black")).add_to(mapa)   
        elif row['city']  == "Bellevue":
            folium.Marker([row["latitud_"], row['longitud_']],
                        radius=10,
                        icon=folium.Icon(icon = "cloud", color = "yellow")).add_to(mapa)

        elif row['city']  == "Redmond":
                folium.Marker([row["latitud_"], row['longitud_']],
                        radius=10,
                        icon=folium.Icon(icon = "cloud", color = "green")).add_to(mapa)  
        elif row['city']  == "Tacoma":
            folium.Marker([row["latitud_"], row['longitud_']],
                        radius=10,
                        icon=folium.Icon(icon = "cloud", color = "purple")).add_to(mapa)

        elif row['city']  == "Vancouver":
            folium.Marker([row["latitud_"], row['longitud_']],
                        radius=10,
                        icon=folium.Icon(icon = "cloud", color = "blue")).add_to(mapa) 

        elif row['city']  == "Long Beach":
            folium.Marker([row["latitud_"], row['longitud_']],
                        radius=10,
                        icon=folium.Icon(icon = "cloud", color = "grey")).add_to(mapa)

        elif row['city']  == "Pasadena":
             folium.Marker([row["latitud_"], row['longitud_']],
                        radius=10, 
                        icon=folium.Icon(icon = "cloud", color = "pink")).add_to(mapa)  
        elif row['city']  == "Anaheim":
            folium.Marker([row["latitud_"], row['longitud_']],
                        radius=10, 
                        icon=folium.Icon(icon = "cloud", color = "violet")).add_to(mapa)

        elif row['city']  == "Irvine":
                folium.Marker([row["latitud_"], row['longitud_']],
                        radius=10, 
                        icon=folium.Icon(icon = "cloud", color = "plum")).add_to(mapa)
        else:
            folium.Marker([row["latitud_"], row['longitud_']],
                        radius=10, 
                        icon=folium.Icon(icon = "cloud", color = "brown")).add_to(mapa)
    except:
        pass
            


mapa

"""for index, row in df3_limpio.iterrows(): # método para iterar por los dataframes
    if row['city'] == "Olympia":
    
        folium.Marker([row["latitud_"], row['longitud_']],
                    radius=10,
                    icon=folium.Icon(icon = "cloud", color = "red")).add_to(mapa)
    elif row['city']  == "Seattle":
        folium.Marker([row["latitud_"], row['longitud_']],
                    radius=10, 
                    icon=folium.Icon(icon = "cloud", color = "orange")).add_to(mapa)
        
    elif row['city']  == "Des Moines":
        folium.Marker([row["latitud_"], row['longitud_']],
                    radius=10,
                    icon=folium.Icon(icon = "cloud", color = "black")).add_to(mapa)   
    elif row['city']  == "Bellevue":
        folium.Marker([row["latitud_"], row['longitud_']],
                    radius=10,
                    icon=folium.Icon(icon = "cloud", color = "yellow")).add_to(mapa)
        
    elif row['city']  == "Redmond":
            folium.Marker([row["latitud_"], row['longitud_']],
                    radius=10,
                    icon=folium.Icon(icon = "cloud", color = "green")).add_to(mapa)  
    elif row['city']  == "Tacoma":
        folium.Marker([row["latitud_"], row['longitud_']],
                    radius=10,
                    icon=folium.Icon(icon = "cloud", color = "purple")).add_to(mapa)
        
    elif row['city']  == "Vancouver":
        folium.Marker([row["latitud_"], row['longitud_']],
                    radius=10,
                    icon=folium.Icon(icon = "cloud", color = "blue")).add_to(mapa) 
    
    elif row['city']  == "Long Beach":
        folium.Marker([row["latitud_"], row['longitud_']],
                    radius=10,
                    icon=folium.Icon(icon = "cloud", color = "grey")).add_to(mapa)
        
    elif row['city']  == "Pasadena":
         folium.Marker([row["latitud_"], row['longitud_']],
                    radius=10, 
                    icon=folium.Icon(icon = "cloud", color = "pink")).add_to(mapa)  
    elif row['city']  == "Anaheim":
        folium.Marker([row["latitud_"], row['longitud_']],
                    radius=10, 
                    icon=folium.Icon(icon = "cloud", color = "violet")).add_to(mapa)
        
    elif row['city']  == "Irvine":
            folium.Marker([row["latitud_"], row['longitud_']],
                    radius=10, 
                    icon=folium.Icon(icon = "cloud", color = "plum")).add_to(mapa)
    else:
        folium.Marker([row["latitud_"], row['longitud_']],
                    radius=10, 
                    icon=folium.Icon(icon = "cloud", color = "brown")).add_to(mapa)"""

df_line = pd.DataFrame(df3_limpio.groupby([df3_limpio['city']])["price"].mean())

df_line.head(3)

ax = plt.axes() # para poder cambiar el estilo de los ejes

ax.set_facecolor("white") # poner el fondo en blanco

ax.spines['bottom'].set_color('black') # color del eje x
ax.spines['top'].set_color('white') # color del eje de arriba
ax.spines['right'].set_color('white') # color del eje de la derecha
ax.spines['left'].set_color('black') # color del eje y

plt.plot(df_line['city'], # eje x
         df_line["price"], # eje y
         color = "black", # para cambiar el color de la línea
         linewidth = 2, # para cambiar el grosor de la línea
         marker = "o") # para poner los marcadores en nuestra línea

plt.xlabel("ciudad") # para poner etiqueta en el eje x
plt.ylabel("precio") # para poner etiqueta en el eje y


plt.show();

sns.scatterplot(data=df3_limpio, x='índice_poder_adqui', y='price/m2',
                s = 100); # tamaño de los marcadores

### Relación Espacio de vivienda y precio

sns.scatterplot(data=df3_limpio, x='building_space', y='price',
                  hue = "city",
                s = 100); # tamaño de los marcadores

df_pie = pd.DataFrame(df3_limpio.groupby(['city'])
                      ["price/m2"].mean()).reset_index()

df_pie.head()

df_pie_1 = pd.DataFrame(df3_limpio.groupby(['city'])
                      ["índice_poder_adqui"].mean()).reset_index()

df_pie_1.head()

df_pie_1 = pd.DataFrame(df3_limpio.groupby(['city'])
                      ["índice_seguridad"].mean()).reset_index()

df_pie_1.head()

sns.scatterplot(data=df3_limpio, x='state', y='índice_contaminación',
                s = 100); # tamaño de los marcadores

sns.scatterplot(data=df3_limpio, x='índice_contaminación', y='índice_calidad_vida',
                s = 100); # tamaño de los marcadores

sns.scatterplot(data=df3_limpio, x='relación_precio_ingresos_propiedades', y='price',
                s = 100); # tamaño de los marcadores

sns.scatterplot(data=df3_limpio, x='state', y='índice_calidad_vida',
                s = 100); # tamaño de los marcadores

sns.scatterplot(data=df3_limpio, x='sanidad', y='índice_contaminación',
                s = 100); # tamaño de los marcadores

def explorar(df, ciudad):
    city = df[df["city"] == f"{ciudad}"]
    print("los principales estadisticos son: ")
    print(city['price/m2'].max())
    print("-----------")
    

def explorar_min(df, ciudad):
    city = df[df["city"] == f"{ciudad}"]
    print("los principales estadisticos son: ")
    print(city['price/m2'].min())
    print("-----------")

def explorar_v2(df, ciudad):
    city = df[df["city"] == f"{ciudad}"]
    print("los principales estadisticos son: ")
    print df3_limipio().T

explorar(df3_limpio, "Irvine")

explorar_min(df3_limpio, "Irvine")

df3_limpio.groupby(["state", 'city'])["price/m2"].mean()

df3_limpio.groupby(["state", 'city'])["price/m2"].max()

df3_limpio.groupby(["state", 'city'])["price/m2"].min()

df3_limpio.groupby(["state", 'city'])["índice_poder_adqui"].mean().max()

df3_limpio.groupby(["state", 'city'])["price/m2"].mean()

df3_limpio.groupby(['city'])["price/m2"].max()

df3_limpio.groupby(['city'])["price/m2"].describe()

df3_limpio.groupby(["state", 'city'])["price/m2"].min()

df3_limpio.groupby(["state", 'city'])["price"].min()

df3_limpio.groupby(["state", 'city'])["price"].mean().round()

df3_limpio.groupby(["state"])["price"].min()

df3_limpio.groupby(["state"])["price"].max().round(0)

df3_limpio.groupby(["state"])["price"].mean().round(0)

df5 = pd.DataFrame(df3_limpio.groupby(["state", 'city','postal_code'])["price/m2"].mean())
df5

df4 = pd.DataFrame(df3_limpio.groupby(["state", 'city','postal_code'])["price/m2"].max())
df4

df4.shape

df3_limpio.groupby(["state", 'city'])["price/m2"].min()
